Move QuantumScript to core
