+++ pennylane/tape/__init__.py
@@ -16,8 +16,14 @@ This subpackage contains the quantum tape, which tracks, queues, and
 validates quantum operations and measurements.
 """
 
+from pennylane.core.qscript import (
+    QuantumScript,
+    QuantumScriptBatch,
+    QuantumScriptOrBatch,
+    make_qscript,
+)
+
 from .operation_recorder import OperationRecorder
-from .qscript import QuantumScript, QuantumScriptBatch, QuantumScriptOrBatch, make_qscript
 from .tape import QuantumTape, QuantumTapeBatch
