+++ pennylane/tape/operation_recorder.py
@@ -15,9 +15,8 @@
 This module contains the :class:`OperationRecorder`.
 """
 
+from pennylane.core.qscript import QuantumScript, process_queue
 from pennylane.queuing import AnnotatedQueue, QueuingManager
-
-from .qscript import QuantumScript, process_queue
 
 
 class OperationRecorder(QuantumScript, AnnotatedQueue):
