+++ pennylane/tape/tape.py
@@ -22,12 +22,11 @@ from collections.abc import Sequence
 from threading import RLock
 
 import pennylane as qp
+from pennylane.core.qscript import QuantumScript, process_queue
 from pennylane.exceptions import PennyLaneDeprecationWarning, QuantumFunctionError
 from pennylane.measurements import CountsMP, ProbabilityMP, SampleMP
 from pennylane.pytrees import register_pytree
 from pennylane.queuing import AnnotatedQueue, QueuingManager
-
-from .qscript import QuantumScript, process_queue
 
 
 def _err_msg_for_some_meas_not_qwc(measurements):
