+++ tach.toml
@@ -95,6 +95,15 @@ expose = ["Shots", "ShotCopies", "ShotsLike"]
 from = ["pennylane.core.shots"]
 
 [[modules]]
+path = "pennylane.core.qscript"
+layer = "core"
+depends_on = ["pennylane.core.operator", "pennylane.core.shots", "pennylane.core.measurements", "pennylane.exceptions", "pennylane.pytrees", "pennylane.queuing", "pennylane.typing", "pennylane.wires"]
+
+[[interfaces]]
+expose = ["QuantumScript", "QuantumScriptBatch", "QuantumScriptOrBatch", "make_qscript", "process_queue"]
+from = ["pennylane.core.qscript"]
+
+[[modules]]
 path = "pennylane.math"
 depends_on = ["pennylane._grad", "pennylane.numpy"]
 cannot_depend_on = ["pennylane.ftqc", "pennylane.labs"]
